fix expr boolean operators evaluation
This commit includes two fixes for expr && and || operators:
* || operator was interpreted as &&
* operands were not evaluated

The || operator being interpreted as && seems like a typo,
probably from copy-pasteing the && logic.

Not evaluating the operands seems like an oversight.
It mean that a && b would be interpreted as literally a && b.
it would not substitute the values of a and b.