in expr, fail immediately when '(' follows an expr
While parsing an expression the parser had a case for '(' appearning
after an expression.

__AFAIK__ this can't occur legally,
the expression would need to look like 1+2(3),
which is not valid in the shell expression language.

I think the branch was only there becuase Precedence::from_token _could_
return Some for Token::LeftParen, so the match statement should handle that
case, but its better to fail there instead.
Otherwise, the parser would fail a bit later with an innacurate error.