add function call in pipeline test for exec
The test case creates a function that writes to stdout using 3 commands.
Then, it uses a pipe to redirect an call to the function to a file.

The purpose of this test is mostly to check function calls.
At the moment there's no return/exit statement,
so verify the function was ran without writing to a file is difficult.