set argv0 in spawn tests, set child process env
The argv[0] should be set to the command used to invoke the executable.
In the tests for job::spawn the first arg was not set to the command.
This led to the first argument, which was
intended for the program, to instead be interpreted as the command.
Coincidentally, this didn't cause any issues,
either because of the commands used in the tests,
(printf "%s" "hello world" is the same as printf "hello world")
or because execve was used instead of execv.

execve was used to wipe the childs environment,
it was only for debugging and comitted by accident.
Forgetting to pass argv0 would have been more obvious if the
environment wasn't empty, because it would have been interpreted as
the arguments to a command.
This would have been especially noticeable on the test that execs rev,
because rev takes a list of files as arguments,
it would try to open every environment variable.