stop storing args and exe in ProcessOptions
In the initial implementation `ProcessOptions` store the exe path
and program args. It makes more sense to keep those seperate from the
program environment since the environment can be applied to multiple
command invocations.

This commit removes the `executable` and `args` fields from
`ProcessOptions`, and makes them paramaters to `ProcessOptions::spawn`.
Default is also now implement for `ProcessOptions`, since it doesn't
need to be initialized with an executable anymore.