Use owned self for ProcessOptions builder methods
Using owned self parameters to ProcessOptions builder methods enables
a copy-on-write pattern when passing around ProcessOptions.
This commit also derives Clone and Debug on ProcessOptions,
because Debug is generally useful, and Clone is needed for
creating a new owned instance of ProcessOptions.

Now, references to ProcessOptions can be passed around, until a function
needs to modify them. Then, it can then clone() and immediately
call builder methods to update the options.